requirements.txt
line-bot-sdk==3.17.1
python-dotenv==1.1.0
Flask==3.1.1
gunicorn==21.2.0 